changed the architecture
